+++ nest-backend/src/users/dto/create-user.dto.ts
@@ -3,3 +3,12 @@ export class CreateUserDto {
   email: string;
   password: string;
 }
+
+export type CreateUserResponse = {
+  message: string;
+  user: {
+    id: string,
+    username: string,
+    email: string,
+  };
+};
