+++ nest-backend/src/users/dto/follow-user.dto.ts
@@ -2,3 +2,10 @@ export class FollowUserDto {
     followerId: string;
     followeeId: string;
 }
+
+export class FollowUserResponse {
+    id: string;
+    followerId: string;
+    followeeId: string;
+    createdAt: Date;
+}
