+++ nest-backend/src/users/dto/get-followees.dto.ts
@@ -1,0 +1,5 @@
+export type GetFolloweesResponse = {
+    id: string,
+    username: string,
+    followingSince: Date,
+};
