+++ nest-backend/src/users/dto/get-followers.dto.ts
@@ -1,0 +1,7 @@
+// ~/user/followers/{userid};
+
+export type GetFollowersResponse = {
+    id: string,
+    username: string,
+    followingSince: Date,
+};
